# server/node/swagger.json
{"definitions":{},"host":"node.ik2.co:5000","info":{"description":"Specification JSONs: [v2](/api-spec/v2), [v3](/api-spec/v3).","title":"server","version":"1.0.0","license":{"name":"ISC"}},"paths":{"/companies":{"get":{"summary":"/companies","consumes":["application/json"],"parameters":[],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"number"},"companyName":{"type":"string"},"website":{"type":"string"},"contacts":{"type":"array","items":{}}},"required":["id","companyName","website","contacts"]}}},"304":{}},"tags":[],"produces":["application/json"]},"post":{"summary":"/companies","consumes":["application/json"],"parameters":[],"responses":{"201":{"description":"Created","schema":{"type":"object","properties":{"id":{"type":"number","example":1695831694459},"companyName":{"type":"string","example":"s"},"website":{"type":"string","example":"s"},"contacts":{"type":"array","items":{},"example":[null]}}}}},"tags":[],"produces":["application/json"]}},"/companies/{id}":{"get":{"summary":"/companies/{id}","consumes":["application/json"],"parameters":[{"name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"type":"object","properties":{"id":{"type":"number","example":1695831825353},"companyName":{"type":"string","example":"d"},"website":{"type":"string","example":"f"},"contacts":{"type":"array","items":{},"example":[null]}}}},"404":{"description":"Not Found","schema":{"type":"string","example":"company not found"}}},"tags":[],"produces":["text/html","application/json"]},"put":{"summary":"/companies/{id}","consumes":["application/json"],"parameters":[{"name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"type":"object","properties":{"id":{"type":"number","example":1695831900389},"companyName":{"type":"string","example":"qw"},"website":{"type":"string","example":"qw"},"contacts":{"type":"array","items":{},"example":[null]}}}},"404":{"description":"Not Found","schema":{"type":"string","example":"company not found"}}},"tags":[],"produces":["application/json","text/html"]},"delete":{"summary":"/companies/{id}","consumes":["application/json"],"parameters":[{"name":"id","in":"path","required":true}],"responses":{"204":{}},"tags":[]}},"/companies/{companyid}/contacts":{"get":{"summary":"/companies/{companyid}/contacts","consumes":["application/json"],"parameters":[{"name":"companyid","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"email":{"type":"string"},"mobile":{"type":"string"},"status":{"type":"string"},"interactions":{"type":"array","items":{"type":"object","properties":{"id":{"type":"number"},"date":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"}}}}}}}},"304":{},"404":{"description":"Not Found","schema":{"type":"string","example":"company not found"}}},"tags":[],"produces":["text/html","application/json"]},"post":{"summary":"/companies/{companyid}/contacts","consumes":["application/json"],"parameters":[{"name":"companyid","in":"path","required":true}],"responses":{"201":{"description":"Created","schema":{"type":"object","properties":{"id":{"type":"number","example":1695831909126},"name":{"type":"string","example":"e"},"email":{"type":"string","example":"e"},"mobile":{"type":"string","example":"e"},"status":{"type":"string","example":"e"},"interactions":{"type":"array","items":{},"example":[null]}}}},"404":{"description":"Not Found","schema":{"type":"string","example":"company not found"}}},"tags":[],"produces":["text/html","application/json"]}},"/companies/{companyid}/contacts/{contactid}":{"get":{"summary":"/companies/{companyid}/contacts/{contactid}","consumes":["application/json"],"parameters":[{"name":"companyid","in":"path","required":true},{"name":"contactid","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"type":"object","properties":{"id":{"type":"number","example":1695832387630},"name":{"type":"string","example":"b"},"email":{"type":"string","example":"b"},"mobile":{"type":"string","example":"b"},"status":{"type":"string","example":"b"},"interactions":{"type":"array","items":{"type":"object","properties":{"id":{"type":"number"},"date":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"}}},"example":[{"id":1695832392060,"date":"c","description":"c","status":"c"}]}}}},"404":{"description":"Not Found","schema":{"type":"string","example":"contact not found"}}},"tags":[],"produces":["text/html","application/json"]},"put":{"summary":"/companies/{companyid}/contacts/{contactid}","consumes":["application/json"],"parameters":[{"name":"companyid","in":"path","required":true},{"name":"contactid","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"type":"object","properties":{"id":{"type":"number","example":1695831909126},"name":{"type":"string","example":"er"},"email":{"type":"string","example":"er"},"mobile":{"type":"string","example":"er"},"status":{"type":"string","example":"er"},"interactions":{"type":"array","items":{},"example":[null]}}}}},"tags":[],"produces":["application/json"]},"delete":{"summary":"/companies/{companyid}/contacts/{contactid}","consumes":["application/json"],"parameters":[{"name":"companyid","in":"path","required":true},{"name":"contactid","in":"path","required":true}],"responses":{"204":{},"404":{"description":"Not Found","schema":{"type":"string","example":"company not found"}}},"tags":[],"produces":["text/html"]}},"/companies/{companyid}/contacts/{contactid}/interactions":{"get":{"summary":"/companies/{companyid}/contacts/{contactid}/interactions","consumes":["application/json"],"parameters":[{"name":"companyid","in":"path","required":true},{"name":"contactid","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"number"},"date":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"}}}}},"304":{}},"tags":[],"produces":["application/json"]},"post":{"summary":"/companies/{companyid}/contacts/{contactid}/interactions","consumes":["application/json"],"parameters":[{"name":"companyid","in":"path","required":true},{"name":"contactid","in":"path","required":true}],"responses":{"201":{"description":"Created","schema":{"type":"object","properties":{"id":{"type":"number","example":1695831918949},"date":{"type":"string","example":"t"},"description":{"type":"string","example":"t"},"status":{"type":"string","example":"t"}}}}},"tags":[],"produces":["application/json"]}},"/companies/{companyid}/contacts/{contactid}/interactions/{interactionid}":{"get":{"summary":"/companies/{companyid}/contacts/{contactid}/interactions/{interactionid}","consumes":["application/json"],"parameters":[{"name":"companyid","in":"path","required":true},{"name":"contactid","in":"path","required":true},{"name":"interactionid","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"type":"object","properties":{"id":{"type":"number","example":1695832392060},"date":{"type":"string","example":"c"},"description":{"type":"string","example":"c"},"status":{"type":"string","example":"c"}}}}},"tags":[],"produces":["application/json"]},"put":{"summary":"/companies/{companyid}/contacts/{contactid}/interactions/{interactionid}","consumes":["application/json"],"parameters":[{"name":"companyid","in":"path","required":true},{"name":"contactid","in":"path","required":true},{"name":"interactionid","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"type":"object","properties":{"id":{"type":"number","example":1695831918949},"date":{"type":"string","example":"ty"},"description":{"type":"string","example":"ty"},"status":{"type":"string","example":"ty"}}}}},"tags":[],"produces":["application/json"]},"delete":{"summary":"/companies/{companyid}/contacts/{contactid}/interactions/{interactionid}","consumes":["application/json"],"parameters":[{"name":"companyid","in":"path","required":true},{"name":"contactid","in":"path","required":true},{"name":"interactionid","in":"path","required":true}],"responses":{"204":{},"404":{"description":"Not Found","schema":{"type":"string","example":"interaction not found"}}},"tags":[],"produces":["text/html"]}},"/*":{"get":{"summary":"/*","consumes":["application/json"],"parameters":[],"responses":{"404":{"description":"Not Found","schema":{"type":"string","example":"Not found"}}},"tags":[],"produces":["text/html"]}},"*":{"get":{"summary":"*","consumes":["application/json"],"parameters":[],"responses":{},"tags":[]}}},"schemes":["http"],"swagger":"2.0","tags":[]}